QOpenGLTextureHelper: fix 3D texture support on ES2
The code first tried to check for the GL_OES_texture_3D extension,
and if present it resolved the pointers to the gl*Tex*3DOES functions.
But then it overwrote those pointers by attempting to resolve the
Desktop GL gl*Tex*3D functions, thus making them unusable.

So, if the extension is found, don't overwrite the pointers.
If the extension is NOT found, keep the general behavior of still
trying to resolve those functions.

(That is going to fail, but refactoring the general behavior
belongs to a separate commit.)

Change-Id: Idaba122cf9500f136e3f79284d3c82284257036d
Reviewed-by: James Turner <james.turner@kdab.com>
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>