Cleanup SPDY stream closing
We should not accept or process messages on closed streams, and unless
we are in a half-closed state (having initiated close ourselves), we
should respond to FIN with a FIN of our own.

This patch means we no longer trigger all the corner case teardown on
common sites that were fixed in earlier patches.

Change-Id: I0d2bab62700a0022a959e66c7053afbad07a9f7e
Reviewed-by: Peter Hartmann <phartmann@blackberry.com>